Move make_user helpers to utils.py
